Creating Registries namespace and moving step and hook registries into it
