+++ integration-test/RefactorProcessorTests.cs
@@ -7,6 +7,7 @@
 
 using Gauge.Dotnet.Models;
 using Gauge.Dotnet.Processors;
+using Gauge.Dotnet.Registries;
 using Gauge.Messages;
 
 namespace Gauge.Dotnet.IntegrationTests
