+++ src/ExecutableRunnerServiceHandler.cs
@@ -8,7 +8,7 @@
 using Gauge.Dotnet.Executors;
 using Gauge.Dotnet.Extensions;
 using Gauge.Dotnet.Loaders;
-using Gauge.Dotnet.Models;
+using Gauge.Dotnet.Registries;
 using Gauge.Messages;
 using Grpc.Core;
