+++ src/Loaders/IStaticLoader.cs
@@ -5,7 +5,7 @@
  *----------------------------------------------------------------*/
 
 
-using Gauge.Dotnet.Models;
+using Gauge.Dotnet.Registries;
 
 namespace Gauge.Dotnet.Loaders
 {
