+++ src/Processors/StepNameProcessor.cs
@@ -6,7 +6,7 @@
 
 
 
-using Gauge.Dotnet.Models;
+using Gauge.Dotnet.Registries;
 using Gauge.Messages;
 
 namespace Gauge.Dotnet.Processors;
