+++ src/Processors/StepNamesProcessor.cs
@@ -5,7 +5,7 @@
  *----------------------------------------------------------------*/
 
 
-using Gauge.Dotnet.Models;
+using Gauge.Dotnet.Registries;
 using Gauge.Messages;
 
 namespace Gauge.Dotnet.Processors;
