+++ src/Registries/IStepRegistry.cs
@@ -4,9 +4,10 @@
  *  See LICENSE.txt in the project root for license information.
  *----------------------------------------------------------------*/
 
+using Gauge.Dotnet.Models;
 using static Gauge.Messages.StepPositionsResponse.Types;
 
-namespace Gauge.Dotnet.Models;
+namespace Gauge.Dotnet.Registries;
 
 public interface IStepRegistry
 {
