+++ test/HookExecutorTests.cs
@@ -9,6 +9,7 @@ using Gauge.CSharp.Lib;
 using Gauge.Dotnet.Executors;
 using Gauge.Dotnet.Loaders;
 using Gauge.Dotnet.Models;
+using Gauge.Dotnet.Registries;
 using Gauge.Dotnet.Strategy;
 using Gauge.Dotnet.UnitTests.Helpers;
 using Gauge.Messages;
