+++ test/HookRegistryTests.cs
@@ -7,7 +7,7 @@
 
 using System.Reflection;
 using Gauge.Dotnet.Loaders;
-using Gauge.Dotnet.Models;
+using Gauge.Dotnet.Registries;
 using Gauge.Dotnet.UnitTests.Helpers;
 using static Gauge.Dotnet.Constants;
