+++ test/Processors/ExecuteStepProcessorTests.cs
@@ -9,6 +9,7 @@ using Gauge.Dotnet.Executors;
 using Gauge.Dotnet.Loaders;
 using Gauge.Dotnet.Models;
 using Gauge.Dotnet.Processors;
+using Gauge.Dotnet.Registries;
 using Gauge.Messages;
 using static Gauge.Dotnet.Constants;
