+++ test/Processors/StepNameProcessorTest.cs
@@ -1,5 +1,6 @@
 ﻿using Gauge.Dotnet.Models;
 using Gauge.Dotnet.Processors;
+using Gauge.Dotnet.Registries;
 using Gauge.Messages;
 
 namespace Gauge.Dotnet.UnitTests.Processors;
