+++ test/Processors/StepNamesProcessorTests.cs
@@ -1,5 +1,5 @@
-﻿using Gauge.Dotnet.Models;
-using Gauge.Dotnet.Processors;
+﻿using Gauge.Dotnet.Processors;
+using Gauge.Dotnet.Registries;
 using Gauge.Messages;
 
 namespace Gauge.Dotnet.UnitTests.Processors;
